// src/apis-client/models/ContainerWithProfile.ts
export interface ContainerWithProfile {
    /**
     * ISO code, primary key, natural key
     * @type {string}
     * @memberof ContainerWithProfile
     */
    containerId: string;
    /**
     * container operational status
     * @type {string}
     * @memberof ContainerWithProfile
     */
    characteristics_size: ContainerWithProfileCharacteristicsSizeEnum;
    /**
     * container type
     * @type {string}
     * @memberof ContainerWithProfile
     */
    characteristics_type: ContainerWithProfileCharacteristicsTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    containerVersion: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    characteristics_maxPayload: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    characteristics_builtYear: string;
    /**
     * updatedAt
     * @type {Date}
     * @memberof ContainerWithProfile
     */
    updatedAt: Date;
    /**
     * createdAt
     * @type {Date}
     * @memberof ContainerWithProfile
     */
    createdAt: Date;
    /**
     * volume
     * @type {number}
     * @memberof ContainerWithProfile
     */
    volume: number;
    /**
     * weight
     * @type {number}
     * @memberof ContainerWithProfile
     */
    characteristics_tare: number;
    /**
     * Inspection Date
     * @type {Date}
     * @memberof ContainerWithProfile
     */
    inspectionDate?: Date;
    /**
     * CSC Cert No.
     * @type {string}
     * @memberof ContainerWithProfile
     */
    cscCertNo?: string;
    /**
     * createdAt
     * @type {Date}
     * @memberof ContainerWithProfile
     */
    cscCertDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    characteristics_grade?: string;
    /**
     * cscCertLink
     * @type {string}
     * @memberof ContainerWithProfile
     */
    cscCertLink?: string;
    /**
     * end panel
     * @type {boolean}
     * @memberof ContainerWithProfile
     */
    electronicsCharging?: boolean;
    /**
     * insurance
     * @type {string}
     * @memberof ContainerWithProfile
     */
    insurance?: string;
    /**
     * Insurance date
     * @type {Date}
     * @memberof ContainerWithProfile
     */
    insuranceDate?: Date;
    /**
     * cameraExt
     * @type {boolean}
     * @memberof ContainerWithProfile
     */
    cameraExt?: boolean;
    /**
     * cameraInt
     * @type {boolean}
     * @memberof ContainerWithProfile
     */
    cameraInt?: boolean;
    /**
     * user id
     * @type {string}
     * @memberof ContainerWithProfile
     */
    lastChangedBy?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    floorType?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    doorType?: string;
    /**
     * the last user who changed the object
     * @type {string}
     * @memberof ContainerWithProfile
     */
    lastChangedByUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerWithProfile
     */
    lastOperation?: ContainerWithProfileLastOperationEnum;
    /**
     * 
     * @type {Array<ContainerSensorsProfile>}
     * @memberof ContainerWithProfile
     */
    containersensorsprofiles: Array<ContainerSensorsProfile>;
    /**
     * 
     * @type {Array<ContainerPanelProfile>}
     * @memberof ContainerWithProfile
     */
    containerpanelprofiles: Array<ContainerPanelProfile>;
    /**
     * 
     * @type {Array<BoardsProfile>}
     * @memberof ContainerWithProfile
     */
    containerboardpairingprofiles: Array<BoardsProfile>;
}

/**
* @export
* @enum {string}
*/
export enum ContainerWithProfileCharacteristicsSizeEnum {
    _20ft = '20ft',
    _40ft = '40ft',
    _45ft = '45ft'
}/**
* @export
* @enum {string}
*/
export enum ContainerWithProfileCharacteristicsTypeEnum {
    Reefer = 'REEFER',
    Flexitank = 'FLEXITANK',
    Normal = 'NORMAL'
}/**
* @export
* @enum {string}
*/
export enum ContainerWithProfileLastOperationEnum {
    D = 'D',
    C = 'C',
    U = 'U'
}