// src/apis-client/models/ContainerPatchInput.ts
export interface ContainerPatchInput {
    /**
     * container operational status
     * @type {string}
     * @memberof ContainerPatchInput
     */
    characteristics_size?: ContainerPatchInputCharacteristicsSizeEnum;
    /**
     * container type
     * @type {string}
     * @memberof ContainerPatchInput
     */
    characteristics_type?: ContainerPatchInputCharacteristicsTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    containerVersion?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    characteristics_maxPayload?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    characteristics_builtYear?: string;
    /**
     * volumne
     * @type {number}
     * @memberof ContainerPatchInput
     */
    volume?: number;
    /**
     * weight
     * @type {number}
     * @memberof ContainerPatchInput
     */
    characteristics_tare?: number;
    /**
     * Inspection Date
     * @type {Date}
     * @memberof ContainerPatchInput
     */
    inspectionDate?: Date;
    /**
     * CSC Cert No.
     * @type {string}
     * @memberof ContainerPatchInput
     */
    cscCertNo?: string;
    /**
     * createdAt
     * @type {Date}
     * @memberof ContainerPatchInput
     */
    cscCertDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    characteristics_grade?: string;
    /**
     * cscCertLink
     * @type {string}
     * @memberof ContainerPatchInput
     */
    cscCertLink?: string;
    /**
     * end panel
     * @type {boolean}
     * @memberof ContainerPatchInput
     */
    electronicsCharging?: boolean;
    /**
     * insurance
     * @type {string}
     * @memberof ContainerPatchInput
     */
    insurance?: string;
    /**
     * Insurance date
     * @type {Date}
     * @memberof ContainerPatchInput
     */
    insuranceDate?: Date;
    /**
     * cameraExt
     * @type {boolean}
     * @memberof ContainerPatchInput
     */
    cameraExt?: boolean;
    /**
     * cameraInt
     * @type {boolean}
     * @memberof ContainerPatchInput
     */
    cameraInt?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    floorType?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    doorType?: string;
    /**
     * 
     * @type {string}
     * @memberof ContainerPatchInput
     */
    depotId?: string;
    /**
     * user id
     * @type {string}
     * @memberof ContainerPatchInput
     */
    lastChangedBy?: string;
}

/**
* @export
* @enum {string}
*/
export enum ContainerPatchInputCharacteristicsSizeEnum {
    _20ft = '20ft',
    _40ft = '40ft',
    _45ft = '45ft'
}/**
* @export
* @enum {string}
*/
export enum ContainerPatchInputCharacteristicsTypeEnum {
    Reefer = 'REEFER',
    Flexitank = 'FLEXITANK',
    Normal = 'NORMAL'
}